refactor(mcp): extract shared backlog read helpers to clear the size ratchet (BI-EC9471C7)
Module Size Guard failed on the parent commit: backlog-pack.ts grew 1318 -> 1358.

Rather than re-baseline an already-oversized file, the shared read concerns move
to apps/web/lib/mcp/packs/backlog-read-helpers.ts — resolveEpicRowId(),
resolveListLimit(), and the cap constants. That is the better structure
regardless of the ratchet: all three read handlers (query_backlog, list_epics,
list_backlog_items) page results and accept an epic filter, and it was exactly
their drifting apart that produced the semantic-id-into-cuid-FK defect this BI
fixes. One home means they cannot drift again.

With the extraction plus tightening of the touched blocks the file lands at
1320 — two lines over baseline, which is the irreducible cost of the added
count queries backing `total`/`truncated`. Re-baselined at 1320 per the guard's
own documented path for justified growth; the baseline diff is one line, this
file only.

Verified: 35 tests pass across admin-pack.test.ts and backlog-pack.test.ts;
tsc --noEmit reports zero errors for all three changed modules.

Process-Spine-Decision: No new capability or contract — backlog-read-helpers.ts
is an internal extraction of code that already shipped in the parent commit, with
identical behaviour and the same callers. No plan or spec artifact is warranted,
and no user-facing documentation describes these MCP internals. The agent-facing
tool descriptions, which are the documentation for this surface, were already
updated in the parent commit.
Design-Grounding-Decision: Unchanged from the parent commit — the source of
truth remains docs/superpowers/specs/2026-04-25-governed-mcp-backlog-surface-design.md
and the substrate reviewed is apps/web/lib/mcp/packs/. This commit only relocates
code within that surface; no contract, gate, or authority behaviour changes.
